Added delete function/buttons and fixed reloadData
added/implemented the ability to add/remove created photos from the app (But are still saved to the gallery), a new button to go along with this function.

Also needed to change the custom CollectionViewLayout for edits since it would only calculate the layout once.  Rather, I added the option of checking whether the number of layouts for cells matched the number of elements in the collectionview, and if not then to recalculate the new datasource of cells.