Minor text placement change/layout fix
Added a bit where if the text context was minimal, to shift the unused context size so the text would align better to the bottom of the template

Also hotfixed a bug in the logic, where I would be comparing layout params and collectionview cell counts and reload/recalculate layouts if so.  There was a small edge case where if a template was removed, and another was added, these counts would remain equal to each other and thus the recalculation would never happen even though there were different templates.